debug: relocate transaction fee logic
